HC UI TEST FIX
Arreglados los UI Tests de HC despues de los cambios en la interfaz y añadidos a un paquete individual.
